Make cwrap Python3 compatible
In python 3, file streams are buffered on multiple levels which means
that it is no longer safe to extract a FILE* pointer from a python file
object. The old CFILE implementation can thus no longer be used. In
order to support function bindings that rely on the old python file to
FILE* mechanism, a thin compatibility layer is introduced.

char* strings from C can no longer be directly used as python str
objects. We therefor need to encode and decode str objects when passed
as arguments to or returned from foreign function calls.

The MethodType constructor in python 3 does not take a class as
argument. This is solved by deferring the construction to the functions
__get__ invocation, where we have access to the calling instance.