Use ctypes.PyDLL instead of CDLL
Doing so will allow us to have the .so live as a Python module. This allows the
.so to export Python names and throw Python exceptions. However, using PyDLL
doesn't unlock the GIL when going to C functions. This could cause performance
problems or even deadlocks in applications that don't expect this.

https://docs.python.org/2/library/ctypes.html#ctypes.PyDLL

We could have this be an opt-in feature by checking if the library contains some
special symbol, eg:

```python
dll = ctypes.CDLL( ... )
if hasattr(dll, "__cwrap_python_module__"):
  dll = ctypes.PyDLL( ... )
```